+++ modules/mauth-authenticator-scala/src/test/scala/com/mdsol/mauth/RequestAuthenticatorBaseSpec.scala
@@ -188,4 +188,36 @@ trait RequestAuthenticatorBaseSpec extends AnyFlatSpec with BeforeAndAfterAll wi
       .build
   }
 
+  def getRequestWithStreamBodyV1: MAuthRequest = {
+    MAuthRequest.Builder.get
+      .withAuthenticationHeaderValue(CLIENT_REQUEST_AUTHENTICATION_BINARY_HEADER_V1)
+      .withTimeHeaderValue(CLIENT_REQUEST_BINARY_TIME_HEADER_VALUE)
+      .withHttpMethod(TestFixtures.REQUEST_METHOD_V2)
+      .withBodyInputStream(new java.io.ByteArrayInputStream(TestFixtures.BINARY_FILE_BODY))
+      .withResourcePath(TestFixtures.REQUEST_PATH_V2)
+      .withQueryParameters(TestFixtures.REQUEST_QUERY_PARAMETERS_V2)
+      .build
+  }
+
+  def getRequestWithStreamBodyV2: MAuthRequest = {
+    MAuthRequest.Builder.get
+      .withAuthenticationHeaderValue(CLIENT_REQUEST_AUTHENTICATION_BINARY_HEADER_V2)
+      .withTimeHeaderValue(CLIENT_REQUEST_BINARY_TIME_HEADER_VALUE)
+      .withHttpMethod(TestFixtures.REQUEST_METHOD_V2)
+      .withBodyInputStream(new java.io.ByteArrayInputStream(TestFixtures.BINARY_FILE_BODY))
+      .withResourcePath(TestFixtures.REQUEST_PATH_V2)
+      .withQueryParameters(TestFixtures.REQUEST_QUERY_PARAMETERS_V2)
+      .build
+  }
+
+  def getRequestWithStreamBodyAndWrongV2Signature: MAuthRequest = {
+    MAuthRequest.Builder.get
+      .withHttpMethod(CLIENT_REQUEST_METHOD)
+      .withMauthHeaders(CLIENT_REQUEST_HEADERS2)
+      .withBodyInputStream(new java.io.ByteArrayInputStream(CLIENT_REQUEST_BODY.getBytes(StandardCharsets.UTF_8)))
+      .withResourcePath(CLIENT_REQUEST_PATH)
+      .withQueryParameters("")
+      .build
+  }
+
 }